#256 Move the api dir in the frontend dir. Other clean ups.
