+++ frontend/requirements.txt
@@ -1,6 +1,7 @@
 Django==1.9.5
 django-cors-headers==1.1.0
 django-tastypie==0.13.3
+MySQL-python==1.2.5
 pytz==2016.6.1
 celery==3.1.23
 pika==0.10.0
